requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-qdrant
qdrant-client
python-dotenv
google-generativeai